рефакторинг
немножко рефакторина - выдернул функции обработки фолловеров в отдельный файлик
