+++ src/act.other.cpp
@@ -14,42 +14,43 @@
 
 #include "act.other.hpp"
 
-#include "world.objects.hpp"
-#include "object.prototypes.hpp"
-#include "logger.hpp"
-#include "obj.hpp"
-#include "comm.h"
-#include "interpreter.h"
-#include "handler.h"
-#include "db.h"
-#include "spells.h"
-#include "skills.h"
-#include "screen.h"
-#include "house.h"
-#include "constants.h"
-#include "dg_scripts.h"
-#include "fightsystem/pk.h"
-#include "fightsystem/fight.h"
-#include "fightsystem/fight_hit.hpp"
-#include "magic.h"
-#include "features.hpp"
-#include "depot.hpp"
-#include "privilege.hpp"
-#include "random.hpp"
+#include "char_obj_utils.inl"
 #include "chars/char.hpp"
 #include "chars/char_player.hpp"
+#include "cmd/follow.h"
+#include "comm.h"
+#include "conf.h"
+#include "constants.h"
+#include "db.h"
+#include "depot.hpp"
+#include "dg_scripts.h"
+#include "features.hpp"
+#include "fightsystem/fight.h"
+#include "fightsystem/fight_hit.hpp"
+#include "fightsystem/pk.h"
+#include "handler.h"
+#include "house.h"
+#include "interpreter.h"
+#include "logger.hpp"
+#include "magic.h"
+#include "msdp.constants.hpp"
+#include "noob.hpp"
+#include "obj.hpp"
+#include "object.prototypes.hpp"
+#include "objsave.h"
+#include "privilege.hpp"
+#include "random.hpp"
 #include "remember.hpp"
 #include "room.hpp"
-#include "objsave.h"
+#include "screen.h"
 #include "shop_ext.hpp"
-#include "noob.hpp"
+#include "skills.h"
+#include "spells.h"
 #include "structs.h"
-#include "conf.h"
 #include "sysdep.h"
-#include "char_obj_utils.inl"
-#include "msdp.constants.hpp"
-#include <sys/stat.h>
+#include "world.objects.hpp"
 
+#include <sys/stat.h>
 #include <sstream>
 #include <fstream>
 #include <string>
