+++ src/act.wizard.cpp
@@ -15,66 +15,67 @@
 #include "act.wizard.hpp"
 
 #include "action.targeting.hpp"
-#include "object.prototypes.hpp"
-#include "world.objects.hpp"
-#include "chars/world.characters.hpp"
-#include "cmd.wiz/stat.h"
-#include "logger.hpp"
-#include "command.shutdown.hpp"
-#include "obj.hpp"
-#include "comm.h"
-#include "interpreter.h"
-#include "handler.h"
-#include "db.h"
-#include "spells.h"
-#include "house.h"
-#include "screen.h"
-#include "skills.h"
-#include "constants.h"
-#include "olc.h"
-#include "dg_scripts.h"
-#include "fightsystem/pk.h"
-#include "im.h"
-#include "top.h"
 #include "ban.hpp"
-#include "description.h"
-#include "title.hpp"
-#include "names.hpp"
-#include "password.hpp"
-#include "privilege.hpp"
-#include "depot.hpp"
-#include "glory.hpp"
-#include "genchar.h"
-#include "file_crc.hpp"
+#include "birth_places.hpp"
+#include "celebrates.hpp"
+#include "char_obj_utils.inl"
 #include "chars/char.hpp"
 #include "chars/char_player.hpp"
-#include "parcel.hpp"
-#include "liquid.hpp"
-#include "modify.h"
-#include "room.hpp"
-#include "glory_misc.hpp"
-#include "glory_const.hpp"
-#include "shop_ext.hpp"
-#include "celebrates.hpp"
 #include "chars/player_races.hpp"
-#include "birth_places.hpp"
-#include "corpse.hpp"
-#include "pugixml.hpp"
-#include "sets_drop.hpp"
-#include "fightsystem/fight.h"
-#include "ext_money.hpp"
-#include "noob.hpp"
-#include "mail.h"
-#include "mob_stat.hpp"
-#include "char_obj_utils.inl"
-#include "utils.h"
-#include "structs.h"
-#include "sysdep.h"
+#include "chars/world.characters.hpp"
+#include "cmd.wiz/stat.h"
+#include "cmd/follow.h"
+#include "comm.h"
+#include "command.shutdown.hpp"
 #include "conf.h"
 #include "config.hpp"
-#include "time_utils.hpp"
+#include "constants.h"
+#include "corpse.hpp"
+#include "db.h"
+#include "depot.hpp"
+#include "description.h"
+#include "dg_scripts.h"
+#include "ext_money.hpp"
+#include "fightsystem/fight.h"
+#include "fightsystem/pk.h"
+#include "file_crc.hpp"
+#include "genchar.h"
 #include "global.objects.hpp"
+#include "glory.hpp"
+#include "glory_const.hpp"
+#include "glory_misc.hpp"
+#include "handler.h"
 #include "heartbeat.hpp"
+#include "house.h"
+#include "im.h"
+#include "interpreter.h"
+#include "liquid.hpp"
+#include "logger.hpp"
+#include "mail.h"
+#include "mob_stat.hpp"
+#include "modify.h"
+#include "names.hpp"
+#include "noob.hpp"
+#include "obj.hpp"
+#include "object.prototypes.hpp"
+#include "olc.h"
+#include "parcel.hpp"
+#include "password.hpp"
+#include "privilege.hpp"
+#include "pugixml.hpp"
+#include "room.hpp"
+#include "screen.h"
+#include "sets_drop.hpp"
+#include "shop_ext.hpp"
+#include "skills.h"
+#include "spells.h"
+#include "structs.h"
+#include "sysdep.h"
+#include "time_utils.hpp"
+#include "title.hpp"
+#include "top.h"
+#include "utils.h"
+#include "world.objects.hpp"
 #include "zone.table.hpp"
 
 #include <boost/format.hpp>
