+++ src/chars/mount.cpp
@@ -2,6 +2,7 @@
 //
 #include "mount.h"
 
+#include "cmd/follow.h"
 #include "handler.h"
 
 void make_horse(CHAR_DATA * horse, CHAR_DATA * ch)
