+++ src/dg_objcmd.cpp
@@ -9,30 +9,31 @@
 *  $Revision$                                                       *
 **************************************************************************/
 
-#include "world.objects.hpp"
-#include "object.prototypes.hpp"
-#include "obj.hpp"
-#include "screen.h"
-#include "dg_scripts.h"
+#include "chars/char.hpp"
+#include "cmd/follow.h"
 #include "comm.h"
-#include "interpreter.h"
-#include "handler.h"
+#include "conf.h"
 #include "db.h"
+#include "dg_scripts.h"
+#include "features.hpp"
+#include "fightsystem/fight.h"
+#include "handler.h"
+#include "im.h"
+#include "interpreter.h"
+#include "logger.hpp"
+#include "magic.h"
+#include "name_list.hpp"
+#include "obj.hpp"
+#include "object.prototypes.hpp"
+#include "room.hpp"
+#include "screen.h"
+#include "skills.h"
 #include "spell_parser.hpp"
 #include "spells.h"
-#include "im.h"
-#include "chars/char.hpp"
-#include "skills.h"
-#include "name_list.hpp"
-#include "room.hpp"
-#include "magic.h"
-#include "fightsystem/fight.h"
-#include "features.hpp"
-#include "logger.hpp"
-#include "utils.h"
 #include "structs.h"
 #include "sysdep.h"
-#include "conf.h"
+#include "utils.h"
+#include "world.objects.hpp"
 
 extern const char *dirs[];
 extern int up_obj_where(OBJ_DATA * obj);
