+++ src/handler.cpp
@@ -14,49 +14,49 @@
 
 #include "handler.h"
 
-#include "chars/world.characters.hpp"
-#include "object.prototypes.hpp"
-#include "world.objects.hpp"
-#include "obj.hpp"
-#include "comm.h"
-#include "db.h"
-#include "glory_const.hpp"
-#include "interpreter.h"
-#include "spells.h"
-#include "skills.h"
-#include "screen.h"
-#include "dg_db_scripts.hpp"
-#include "dg_scripts.h"
 #include "auction.h"
-#include "features.hpp"
-#include "house.h"
-#include "exchange.h"
+#include "backtrace.hpp"
+#include "char_obj_utils.inl"
 #include "chars/char.hpp"
 #include "chars/char_player.hpp"
-#include "liquid.hpp"
-#include "magic.h"
-#include "poison.hpp"
-#include "name_list.hpp"
-#include "room.hpp"
-#include "named_stuff.hpp"
-#include "glory_const.hpp"
+#include "chars/world.characters.hpp"
+#include "cmd/follow.h"
+#include "comm.h"
+#include "conf.h"
+#include "constants.h"
+#include "db.h"
+#include "dg_db_scripts.hpp"
+#include "dg_scripts.h"
+#include "exchange.h"
+#include "ext_money.hpp"
+#include "features.hpp"
 #include "fightsystem/fight.h"
 #include "fightsystem/pk.h"
-#include "ext_money.hpp"
-#include "noob.hpp"
-#include "obj_sets.hpp"
-#include "char_obj_utils.inl"
-#include "constants.h"
-#include "spell_parser.hpp"
+#include "glory_const.hpp"
+#include "glory_const.hpp"
+#include "house.h"
+#include "interpreter.h"
+#include "liquid.hpp"
 #include "logger.hpp"
+#include "magic.h"
+#include "name_list.hpp"
+#include "named_stuff.hpp"
+#include "noob.hpp"
+#include "obj.hpp"
+#include "obj_sets.hpp"
+#include "object.prototypes.hpp"
+#include "poison.hpp"
+#include "room.hpp"
+#include "screen.h"
+#include "skills.h"
+#include "spell_parser.hpp"
+#include "spells.h"
 #include "structs.h"
 #include "sysdep.h"
-#include "conf.h"
+#include "world.objects.hpp"
 #include "zone.table.hpp"
-#include "backtrace.hpp"
 
 #include <math.h>
-
 #include <unordered_set>
 #include <sstream>
