+++ src/interpreter.cpp
@@ -15,77 +15,77 @@
 
 #include "interpreter.h"
 
-#include "cmd/retreat.h"
-#include "cmd/order.h"
-#include "cmd/track.h"
-#include "cmd/telegram.h"
-#include "cmd/hire.h"
-
-#include "cmd.wiz/stat.h"
-
-#include "skills/manadrain.h"
-#include "skills/flee.h"
-#include "skills/bash.h"
-#include "skills/stun.h"
-#include "skills/resque.h"
-#include "skills/kick.h"
-#include "skills/strangle.h"
-#include "skills/chopoff.h"
-#include "skills/disarm.h"
-#include "skills/stupor.h"
-#include "skills/ironwind.h"
-#include "skills/throw.h"
-#include "skills/mighthit.h"
-#include "skills/block.h"
-#include "skills/parry.h"
-#include "skills/protect.h"
-#include "skills/turnundead.h"
-#include "fightsystem/assist.h"
-#include "fightsystem/start.fight.h"
-#include "fightsystem/mobact.hpp"
-#include "cmd/mercenary.h"
 #include "act.movement.hpp"
-#include "chars/world.characters.hpp"
-#include "object.prototypes.hpp"
-#include "logger.hpp"
-#include "craft.commands.hpp"
-#include "heartbeat.commands.hpp"
-#include "obj.hpp"
-#include "comm.h"
-#include "constants.h"
-#include "db.h"
-#include "spells.h"
-#include "skills.h"
-#include "handler.h"
-#include "house.h"
-#include "mail.h"
-#include "screen.h"
-#include "olc.h"
-#include "dg_scripts.h"
-#include "fightsystem/pk.h"
-#include "genchar.h"
 #include "ban.hpp"
-#include "item.creation.hpp"
-#include "features.hpp"
 #include "boards.h"
-#include "top.h"
-#include "title.hpp"
-#include "names.hpp"
-#include "password.hpp"
-#include "privilege.hpp"
-#include "depot.hpp"
-#include "glory.hpp"
 #include "chars/char.hpp"
 #include "chars/char_player.hpp"
-#include "parcel.hpp"
-#include "liquid.hpp"
-#include "name_list.hpp"
-#include "modify.h"
-#include "room.hpp"
+#include "chars/world.characters.hpp"
+#include "cmd.wiz/stat.h"
+#include "cmd/follow.h"
+#include "cmd/hire.h"
+#include "cmd/mercenary.h"
+#include "cmd/order.h"
+#include "cmd/retreat.h"
+#include "cmd/telegram.h"
+#include "cmd/track.h"
+#include "comm.h"
+#include "constants.h"
+#include "craft.commands.hpp"
+#include "db.h"
+#include "depot.hpp"
+#include "dg_scripts.h"
+#include "features.hpp"
+#include "fightsystem/assist.h"
+#include "fightsystem/mobact.hpp"
+#include "fightsystem/pk.h"
+#include "fightsystem/start.fight.h"
+#include "genchar.h"
+#include "glory.hpp"
 #include "glory_const.hpp"
 #include "glory_misc.hpp"
+#include "handler.h"
+#include "heartbeat.commands.hpp"
+#include "house.h"
+#include "item.creation.hpp"
+#include "liquid.hpp"
+#include "logger.hpp"
+#include "mail.h"
+#include "modify.h"
+#include "name_list.hpp"
 #include "named_stuff.hpp"
+#include "names.hpp"
+#include "obj.hpp"
+#include "object.prototypes.hpp"
+#include "olc.h"
+#include "parcel.hpp"
+#include "password.hpp"
+#include "privilege.hpp"
+#include "room.hpp"
+#include "screen.h"
+#include "skills.h"
+#include "skills/bash.h"
+#include "skills/block.h"
+#include "skills/chopoff.h"
+#include "skills/disarm.h"
+#include "skills/flee.h"
+#include "skills/ironwind.h"
+#include "skills/kick.h"
+#include "skills/manadrain.h"
+#include "skills/mighthit.h"
+#include "skills/parry.h"
+#include "skills/protect.h"
+#include "skills/resque.h"
+#include "skills/strangle.h"
+#include "skills/stun.h"
+#include "skills/stupor.h"
+#include "skills/throw.h"
+#include "skills/turnundead.h"
+#include "spells.h"
 #include "time.h"
+#include "title.hpp"
+#include "top.h"
+
 #if defined WITH_SCRIPTING
 #include "scripting.hpp"
 #endif
