+++ src/logger.hpp
@@ -10,6 +10,9 @@
 #include <thread>
 
 extern FILE *logfile;
+extern std::list<FILE *> opened_files;
+
+void pers_log(CHAR_DATA *ch, const char *format, ...) __attribute__((format(printf,2,3)));
 
 void log(const char *format, ...) __attribute__((format(printf, 1, 2)));
 void vlog(const char* format, va_list args) __attribute__((format(printf, 1, 0)));
