+++ src/spells.cpp
@@ -14,45 +14,46 @@
 
 #include "spells.h"
 
-#include "skills/townportal.h"
-#include "cmd/hire.h"
-
-#include "coredump.hpp"
-#include "world.objects.hpp"
-#include "object.prototypes.hpp"
-#include "char_obj_utils.inl"
-#include "obj.hpp"
-#include "comm.h"
-#include "skills.h"
-#include "handler.h"
-#include "db.h"
-#include "constants.h"
-#include "interpreter.h"
-#include "dg_scripts.h"
-#include "screen.h"
-#include "house.h"
-#include "fightsystem/pk.h"
-#include "features.hpp"
-#include "im.h"
-#include "deathtrap.hpp"
-#include "privilege.hpp"
-#include "chars/char.hpp"
-#include "depot.hpp"
-#include "parcel.hpp"
-#include "liquid.hpp"
-#include "modify.h"
-#include "room.hpp"
 #include "birth_places.hpp"
-#include "obj_sets.hpp"
+#include "char_obj_utils.inl"
+#include "chars/char.hpp"
+#include "chars/world.characters.hpp"
+#include "cmd/follow.h"
+#include "cmd/hire.h"
+#include "comm.h"
+#include "conf.h"
+#include "constants.h"
+#include "coredump.hpp"
+#include "db.h"
+#include "deathtrap.hpp"
+#include "depot.hpp"
+#include "dg_scripts.h"
+#include "features.hpp"
+#include "fightsystem/mobact.hpp"
+#include "fightsystem/pk.h"
+#include "handler.h"
+#include "house.h"
+#include "im.h"
+#include "interpreter.h"
+#include "liquid.hpp"
 #include "logger.hpp"
 #include "magic.h"
+#include "modify.h"
+#include "obj.hpp"
+#include "obj_sets.hpp"
+#include "object.prototypes.hpp"
+#include "parcel.hpp"
+#include "privilege.hpp"
+#include "room.hpp"
+#include "screen.h"
+#include "skills.h"
+#include "skills/flee.h"
+#include "skills/townportal.h"
 #include "structs.h"
 #include "sysdep.h"
-#include "conf.h"
-#include "chars/world.characters.hpp"
+#include "world.objects.hpp"
 #include "zone.table.hpp"
-#include "skills/flee.h"
-#include "fightsystem/mobact.hpp"
+
 #include <math.h>
 #include <vector>
