+++ src/utils.cpp
@@ -490,77 +490,6 @@ int strn_cmp(const std::string &arg1, const std::string &arg2, size_t n)
 		return (LOWER(arg1[i]) - LOWER('\0'));
 }
 
-// дескрипторы открытых файлов логов для сброса буфера при креше
-std::list<FILE *> opened_files;
-
-bool no_bad_affects(OBJ_DATA *obj)
-{
-	static std::list<EWeaponAffectFlag> bad_waffects =
-	{
-		EWeaponAffectFlag::WAFF_HOLD,
-		EWeaponAffectFlag::WAFF_SANCTUARY,
-		EWeaponAffectFlag::WAFF_PRISMATIC_AURA,
-		EWeaponAffectFlag::WAFF_POISON,
-		EWeaponAffectFlag::WAFF_SILENCE,
-		EWeaponAffectFlag::WAFF_DEAFNESS,
-		EWeaponAffectFlag::WAFF_HAEMORRAGIA,
-		EWeaponAffectFlag::WAFF_BLINDNESS,
-		EWeaponAffectFlag::WAFF_SLEEP,
-		EWeaponAffectFlag::WAFF_HOLY_DARK
-	};
-	for (const auto wa : bad_waffects)
-	{
-		if (OBJ_AFFECT(obj, wa))
-		{
-			return false;
-		}
-	}
-	return true;
-}
-
-/**
-* Файл персонального лога терь открывается один раз за каждый вход плеера в игру.
-* Дескриптор открытого файла у плеера же и хранится (закрывает при con_close).
-*/
-void pers_log(CHAR_DATA *ch, const char *format, ...)
-{
-	if (!ch)
-	{
-		log("NULL character resieved! (%s %s %d)", __FILE__, __func__, __LINE__);
-		return;
-	}
-
-	if (!format)
-	{
-		format = "SYSERR: pers_log received a NULL format.";
-	}
-
-	if (!ch->desc->pers_log)
-	{
-		char filename[128], name[64], *ptr;
-		strcpy(name, GET_NAME(ch));
-		for (ptr = name; *ptr; ptr++)
-		{
-			*ptr = LOWER(AtoL(*ptr));
-		}
-		sprintf(filename, "../log/perslog/%s.log", name);
-		ch->desc->pers_log = fopen(filename, "a");
-		if (!ch->desc->pers_log)
-		{
-			log("SYSERR: error open %s (%s %s %d)", filename, __FILE__, __func__, __LINE__);
-			return;
-		}
-		opened_files.push_back(ch->desc->pers_log);
-	}
-
-	write_time(ch->desc->pers_log);
-	va_list args;
-	va_start(args, format);
-	vfprintf(ch->desc->pers_log, format, args);
-	va_end(args);
-	fprintf(ch->desc->pers_log, "\n");
-}
-
 // the "touch" command, essentially.
 int touch(const char *path)
 {
@@ -614,8 +543,6 @@ TIME_INFO_DATA *real_time_passed(time_t t2, time_t t1)
 	return (&now);
 }
 
-
-
 // Calculate the MUD time passed over the last t2-t1 centuries (secs) //
 TIME_INFO_DATA *mud_time_passed(time_t t2, time_t t1)
 {
@@ -638,8 +565,6 @@ TIME_INFO_DATA *mud_time_passed(time_t t2, time_t t1)
 	return (&now);
 }
 
-
-
 TIME_INFO_DATA *age(const CHAR_DATA * ch)
 {
 	static TIME_INFO_DATA player_age;
@@ -649,169 +574,6 @@ TIME_INFO_DATA *age(const CHAR_DATA * ch)
 	player_age.year += 17;	// All players start at 17 //
 
 	return (&player_age);
-}
-
-// Check if making CH follow VICTIM will create an illegal //
-// Follow "Loop/circle"                                    //
-bool circle_follow(CHAR_DATA * ch, CHAR_DATA * victim)
-{
-	for (auto k = victim; k; k = k->get_master())
-	{
-		if (k->get_master() == k)
-		{
-			k->set_master(nullptr);
-			return false;
-		}
-
-		if (k == ch)
-		{
-			return true;
-		}
-	}
-
-	return false;
-}
-
-
-// Called when stop following persons, or stopping charm //
-// This will NOT do if a character quits/dies!!          //
-// При возврате 1 использовать ch нельзя, т.к. прошли через extract_char
-// TODO: по всем вызовам не проходил, может еще где-то коряво вызывается, кроме передачи скакунов -- Krodo
-// при персонаже на входе - пуржить не должно полюбому, если начнет, как минимум в change_leader будут глюки
-bool stop_follower(CHAR_DATA * ch, int mode)
-{
-	struct follow_type *j, *k;
-	int i;
-
-	//log("[Stop follower] Start function(%s->%s)",ch ? GET_NAME(ch) : "none",
-	//      ch->master ? GET_NAME(ch->master) : "none");
-
-	if (!ch->has_master())
-	{
-		log("SYSERR: stop_follower(%s) without master", GET_NAME(ch));
-		return (FALSE);
-	}
-
-	// для смены лидера без лишнего спама
-	if (!IS_SET(mode, SF_SILENCE))
-	{
-		act("Вы прекратили следовать за $N4.", FALSE, ch, 0, ch->get_master(), TO_CHAR);
-		act("$n прекратил$g следовать за $N4.", TRUE, ch, 0, ch->get_master(), TO_NOTVICT | TO_ARENA_LISTEN);
-	}
-
-	//log("[Stop follower] Stop horse");
-	if (ch->get_master()->get_horse() == ch && ch->get_master()->ahorse())
-	{
-		ch->drop_from_horse();
-	}
-	else
-	{
-		act("$n прекратил$g следовать за вами.", TRUE, ch, 0, ch->get_master(), TO_VICT);
-	}
-
-	//log("[Stop follower] Remove from followers list");
-	if (!ch->get_master()->followers)
-	{
-		log("[Stop follower] SYSERR: Followers absent for %s (master %s).", GET_NAME(ch), GET_NAME(ch->get_master()));
-	}
-	else if (ch->get_master()->followers->follower == ch)  	// Head of follower-list?
-	{
-		k = ch->get_master()->followers;
-		ch->get_master()->followers = k->next;
-		if (!ch->get_master()->followers
-			&& !ch->get_master()->has_master())
-		{
-			//AFF_FLAGS(ch->get_master()).unset(EAffectFlag::AFF_GROUP);
-			ch->get_master()->removeGroupFlags();
-		}
-		free(k);
-	}
-	else  		// locate follower who is not head of list
-	{
-		for (k = ch->get_master()->followers; k->next && k->next->follower != ch; k = k->next);
-		if (!k->next)
-		{
-			log("[Stop follower] SYSERR: Undefined %s in %s followers list.", GET_NAME(ch), GET_NAME(ch->get_master()));
-		}
-		else
-		{
-			j = k->next;
-			k->next = j->next;
-			free(j);
-		}
-	}
-
-	ch->set_master(nullptr);
-	//AFF_FLAGS(ch).unset(EAffectFlag::AFF_GROUP);
-	ch->removeGroupFlags();
-
-	if (AFF_FLAGGED(ch, EAffectFlag::AFF_CHARM)
-		|| AFF_FLAGGED(ch, EAffectFlag::AFF_HELPER)
-		|| IS_SET(mode, SF_CHARMLOST))
-	{
-		if (affected_by_spell(ch, SPELL_CHARM))
-		{
-			affect_from_char(ch, SPELL_CHARM);
-		}
-		EXTRACT_TIMER(ch) = 5;
-		AFF_FLAGS(ch).unset(EAffectFlag::AFF_CHARM);
-
-		if (ch->get_fighting())
-		{
-			stop_fighting(ch, TRUE);
-		}
-
-		if (IS_NPC(ch))
-		{
-			if (MOB_FLAGGED(ch, MOB_CORPSE))
-			{
-				act("Налетевший ветер развеял $n3, не оставив и следа.", TRUE, ch, 0, 0, TO_ROOM | TO_ARENA_LISTEN);
-				GET_LASTROOM(ch) = GET_ROOM_VNUM(ch->in_room);
-				perform_drop_gold(ch, ch->get_gold());
-				ch->set_gold(0);
-				extract_char(ch, FALSE);
-				return (TRUE);
-			}
-			else if (AFF_FLAGGED(ch, EAffectFlag::AFF_HELPER))
-			{
-				AFF_FLAGS(ch).unset(EAffectFlag::AFF_HELPER);
-			}
-		}
-	}
-
-	if (IS_NPC(ch)
-		&& !MOB_FLAGGED(ch, MOB_PLAYER_SUMMON)	//Не ресетим флаги, если моб призван игроком
-		&& (i = GET_MOB_RNUM(ch)) >= 0)
-	{
-		MOB_FLAGS(ch) = MOB_FLAGS(mob_proto + i);
-	}
-
-	return (FALSE);
-}
-
-
-
-// * Called when a character that follows/is followed dies
-bool die_follower(CHAR_DATA * ch)
-{
-	struct follow_type *j, *k = ch->followers;
-
-	if (ch->has_master() && stop_follower(ch, SF_FOLLOWERDIE))
-	{
-		//  чармиса спуржили в stop_follower
-		return true;
-	}
-
-	if (ch->ahorse()) {
-		AFF_FLAGS(ch).unset(EAffectFlag::AFF_HORSE);
-	}
-
-	for (k = ch->followers; k; k = j)
-	{
-		j = k->next;
-		stop_follower(k->follower, SF_MASTERDIE);
-	}
-	return false;
 }
 
 /*
